Use a tempfile as lilypond output target, then copy to real target
That way we deal with the annoying habit of lilypond to check if the
target is a directory. Since lilypond adds the extension, this could
lead to incorrect behavior, for example when compiling a sub-directory
of the current directory.

The lilypond invoker now generates a temporary file name, gives it to
lilypond as a target, and then copies the output to the user supplied
target.